Fix type casting in ft_memmove loop condition

+++ ft_memmove.c
@@ -26,7 +26,7 @@ void	*ft_memmove(void *dest, const void *src, size_t n)
 		i = n - 1;
 		copy_direction = -1;
 	}
-	while (i < n && i >= 0)
+	while (i < n && (int)i >= 0)
 	{
 		((char *)dest)[i] = ((char *)src)[i];
 		i += copy_direction;
